+++ modules/shared/components/atoms/avatar/Avatar.test.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import * as renderer from 'react-test-renderer';
-import Avatar, { determineAvatarSize } from './Avatar';
+import Avatar from './Avatar';
+import { determineAvatarSize } from './utils';
 
 describe('Avatar', () => {
   it('make sure that imageSize works correctly', () => {
