+++ modules/shared/components/atoms/avatar/Avatar.tsx
@@ -1,24 +1,12 @@
-import type { FC, ReactElement } from 'react';
 import React from 'react';
 import Image from 'next/image';
 import classNames from 'classnames';
+import type { FC, ReactElement } from 'react';
 import type { IAvatar } from './IAvatar';
 import AnonymousIcon from '../../icons/anonymous.svg';
 import NotFilledIcon from '../../icons/notFilled.svg';
 import styles from './Avatar.module.css';
-
-export const determineAvatarSize = (componentSize: string): number => {
-  const smallSize = 32;
-  const mediumSize = 40;
-  const largeSize = 56;
-  if (componentSize === 'small') {
-    return smallSize;
-  }
-  if (componentSize === 'medium') {
-    return mediumSize;
-  }
-  return largeSize;
-};
+import { determineAvatarSize } from './utils';
 
 const Avatar: FC<IAvatar.IProps> = ({
   size,
