+++ modules/shared/components/atoms/avatar/utils.ts
@@ -1,0 +1,12 @@
+export const determineAvatarSize = (componentSize: string): number => {
+  const smallSize = 32;
+  const mediumSize = 40;
+  const largeSize = 56;
+  if (componentSize === 'small') {
+    return smallSize;
+  }
+  if (componentSize === 'medium') {
+    return mediumSize;
+  }
+  return largeSize;
+};
